Handle Ctrl-c more gracefully
When the user stops the script via Ctrl-c, the intention is almost
certainly not to be in some vaguely defined intermediate state with the
original HEAD commit reverted and potentially part of its changes
staged.
With this change we bring some order to this scenario: we install a
handler for SIGINT (sent when the user presses Ctrl-c), which takes care
of resetting the git repository back to the state when we started.