Handle git commit failure more gracefully
A git commit could fail for legitimate reasons, such as the user
clearing the commit message. From the perspective of this script, it is
probably wisest to not propagate up the reported error and fail the
overall operation, as it may just be a user's attempt to redo the last
step or similar.
To that end, with this change we handle such failures more gracefully
and do not exit the script. On top of that, we make sure to undo any