Revert state on regular command failures
At this point we no longer anticipate any command failures, though they
are still conceivable (e.g., git reset HEAD^ could fail in a repository
without a commit).
We generally do not want the user to see any intermediate state if we
were to exit the script due to such a failure. As such, with this change
we register our trap handler that reverts all changes made for command
failures as well.