Add support for splitting out changes to a subset of files
Sometimes a user's intention is to only split out changes to a certain
set of files known very well in advance, but with the current logic said
user would be asked to select hunks on each and every file committed
earlier.
With this change we accept additional positional arguments that allow
the user to only split out changes to a pre-selected list of files.