Improve handling of removed files in commits
When a file is removed by a commit, the script currently just implicitly
includes it in the first commit created. That is undesirable behavior:
clearly the user should have control over which commit to include them
in.
With this change we fix up this short coming, by passing the
'--ignore-removal' to the initial 'git add --intent-to-add'.